Moved ValidatedSpec dependency from Spec to SpecValidator.
