+++ src/mlte/evidence/identifier.py
@@ -3,7 +3,6 @@ Identifier class implementation.
 """
 
 from __future__ import annotations
-from typing import Any
 
 
 class Identifier:
