+++ test/spec/test_specvalidator.py
@@ -6,10 +6,10 @@ import pytest
 
 from mlte.spec import Spec, Requirement, SpecValidator
 from mlte.property.costs import StorageCost
-
+from mlte.validation import Result
 from mlte.value.types import Integer
 from mlte.evidence import EvidenceMetadata
-
+from mlte.measurement.storage import LocalObjectSize
 
 def test_no_property():
     # Spec validator does not have value for property.
@@ -47,3 +47,18 @@ def test_success():
 
     validatedSpec = specValidator.validate()
     assert validatedSpec is not None
+
+def test_no_result():
+    # Spec does not have value for requirement.
+    spec = Spec(
+        {
+            StorageCost("rationale"): [
+                Requirement("test", LocalObjectSize.value().less_than(3))
+            ]
+        }
+    )
+    specValidator = SpecValidator(spec)    
+
+    results: dict[str, Result] = {}
+    with pytest.raises(RuntimeError):
+        _ = specValidator._generate_validatedspec(results)
